backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
langgraph>=0.0.10
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncpg>=0.28.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
loguru>=0.7.0
